Add tests for Position struct and move handling in shogi
- Introduced a new tests module for the shogi engine.
- Implemented comprehensive tests for the Position struct, covering normal moves, captures, promotions, drops, and various edge cases.
- Verified the functionality of do_move and undo_move methods, ensuring reversibility and correct state management.
- Added tests for detecting repetitions and handling null moves.
