+++ packages/rust-core/crates/engine-core/src/shogi/mod.rs
@@ -3,7 +3,10 @@ pub mod board;
 pub mod moves;
 pub mod piece_constants;
 
+#[cfg(test)]
+mod tests;
+
 pub use attacks::{AttackTables, Direction, ATTACK_TABLES};
-pub use board::{Bitboard, Board, Color, Piece, PieceType, Position, Square};
+pub use board::{Bitboard, Board, Color, Piece, PieceType, Position, Square, UndoInfo};
 pub use moves::{Move, MoveList};
 pub use piece_constants::*;
