+++ packages/rust-core/crates/engine-core/src/shogi/tests/mod.rs
@@ -1,0 +1,4 @@
+//! Test modules for shogi board implementation
+
+#[cfg(test)]
+mod position_tests;
